requirements.txt
django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.10,<4.0
channels>=3.0,<4.0
